requirements.txt

fastapi==0.75.0

uvicorn==0.17.4

pydantic[email]

motor==2.5.1

razorpay==1.3.0

python-decouple==3.6

fastapi-mail==1.0.6